refactor: move namespace declarations and fetch_ns() to io.py
